+++ packages/react-components/react-teaching-popover-preview/src/components/TeachingPopoverCarousel/useTeachingPopoverCarouselStyles.styles.ts
@@ -1,35 +1,21 @@
 import { makeStyles, mergeClasses } from '@griffel/react';
-import type { SlotClassNames } from '@fluentui/react-utilities';
 import type { TeachingPopoverCarouselSlots, TeachingPopoverCarouselState } from './TeachingPopoverCarousel.types';
+import type { SlotClassNames } from '@fluentui/react-utilities';
 
 export const teachingPopoverCarouselClassNames: SlotClassNames<TeachingPopoverCarouselSlots> = {
   root: 'fui-TeachingPopoverCarousel',
-  // TODO: add class names for all slots on TeachingPopoverCarouselSlots.
-  // Should be of the form `<slotName>: 'fui-TeachingPopoverCarousel__<slotName>`
 };
 
-/**
- * Styles for the root slot
- */
+// Todo: Page change animation & styles
 const useStyles = makeStyles({
-  root: {
-    // TODO Add default styles for the root element
-  },
-
-  // TODO add additional classes for different states and/or slots
+  root: {},
 });
 
-/**
- * Apply styling to the TeachingPopoverCarousel slots based on the state
- */
-export const useTeachingPopoverCarouselStyles_unstable = (
-  state: TeachingPopoverCarouselState,
-): TeachingPopoverCarouselState => {
+/** Applies style classnames to slots */
+export const useTeachingPopoverCarouselStyles_unstable = (state: TeachingPopoverCarouselState) => {
   const styles = useStyles();
-  state.root.className = mergeClasses(teachingPopoverCarouselClassNames.root, styles.root, state.root.className);
 
-  // TODO Add class names to slots, for example:
-  // state.mySlot.className = mergeClasses(styles.mySlot, state.mySlot.className);
+  state.root.className = mergeClasses(teachingPopoverCarouselClassNames.root, styles.root, state.root.className);
 
   return state;
 };